+++ python/shipDet_conf.py
@@ -750,22 +750,6 @@ def configure(run, ship_geo):
         detectorList.append(Strawtubes)
 
     if ship_geo.EcalOption == 1:  # shashlik design TP
-        if ship_geo.preshowerOption > 0:
-            Preshower = ROOT.preshower("Preshower", ROOT.kTRUE)
-            Preshower.SetZStationPosition2(
-                ship_geo.PreshowerStation0.z, ship_geo.PreshowerStation1.z
-            )
-            Preshower.SetZFilterPosition2(
-                ship_geo.PreshowerFilter0.z, ship_geo.PreshowerFilter1.z
-            )
-            Preshower.SetXMax(ship_geo.Preshower.XMax)
-            Preshower.SetYMax(ship_geo.Preshower.YMax)
-            Preshower.SetActiveThickness(ship_geo.Preshower.ActiveThickness)
-            Preshower.SetFilterThickness2(
-                ship_geo.Preshower.FilterThickness0, ship_geo.Preshower.FilterThickness1
-            )
-            detectorList.append(Preshower)
-
         ecal, EcalZSize = posEcal(ship_geo.ecal.z, ship_geo.ecal.File)
         detectorList.append(ecal)